pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "pptx-components"
version = "0.1.0"
description = "Reusable PowerPoint components for python-pptx slide generation."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = [
    "powerpoint",
    "python-pptx",
    "pptx",
    "components",
    "slides",
    "presentation",
]
dependencies = [
    "python-pptx>=1.0.0",
    "pywin32>=306; platform_system == 'Windows'",
]

[project.optional-dependencies]
test = [
    "pytest>=8",
]
docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.26.0",
]
dev = [
    "black>=24.10.0",
    "pytest>=8",
]

[tool.pytest.ini_options]
testpaths = ["tests", "."]
python_files = ["test_*.py"]

[tool.black]
line-length = 88
